Use the default multiprocessing pool configuration.
Previously we were forcing use of fork for all OSes to try to juice
performance at the cost of running into threading issues at some point
when things got more complex. They may have reached that point with
CI lockups happening on macOS. Revert to the default pool setup to
attempt to fix macOS hangs.

To support this, also propagate any CLI global `Variables` adjustments
to subprocesses. Previously these were only propagated for the current
process; now the adjustments propagate through the environment so that
subprocess see the same adjustments. In particular, this ensures
subprocess use the same `PEX_ROOT` when a fallback has to be used to
work around lack of write perms.
