+++ tests/resolve/test_target_options.py
@@ -10,6 +10,7 @@ from argparse import ArgumentParser, ArgumentTypeError
 import pytest
 
 import pex.resolve.target_configuration
+from pex.common import environment_as
 from pex.interpreter import PythonInterpreter
 from pex.pep_425 import CompatibilityTags
 from pex.pep_508 import MarkerEnvironment
@@ -20,7 +21,7 @@ from pex.resolve.target_configuration import InterpreterConstraintsNotSatisfied
 from pex.targets import CompletePlatform, Targets
 from pex.typing import TYPE_CHECKING
 from pex.variables import ENV
-from testing import IS_MAC, environment_as
+from testing import IS_MAC
 
 if TYPE_CHECKING:
     from typing import Any, Dict, Iterable, List, Optional, Tuple, Type
