+++ tests/test_atomic_directory.py
@@ -9,9 +9,8 @@ from contextlib import contextmanager
 import pytest
 
 from pex.atomic_directory import AtomicDirectory, FileLockStyle, _is_bsd_lock, atomic_directory
-from pex.common import temporary_dir, touch
+from pex.common import environment_as, temporary_dir, touch
 from pex.typing import TYPE_CHECKING
-from testing import environment_as
 
 try:
     from unittest import mock
