+++ tests/test_interpreter.py
@@ -14,7 +14,7 @@ from textwrap import dedent
 import pytest
 
 from pex.cache.dirs import InterpreterDir
-from pex.common import chmod_plus_x, safe_mkdir, safe_mkdtemp, temporary_dir, touch
+from pex.common import chmod_plus_x, environment_as, safe_mkdir, safe_mkdtemp, temporary_dir, touch
 from pex.executor import Executor
 from pex.interpreter import PythonInterpreter, create_shebang
 from pex.jobs import Job
@@ -30,7 +30,6 @@ from testing import (
     ensure_python_distribution,
     ensure_python_interpreter,
     ensure_python_venv,
-    environment_as,
     pushd,
 )
 from testing.pytest.tmp import TempdirFactory
