+++ tests/test_variables.py
@@ -7,12 +7,12 @@ import warnings
 import pytest
 
 from pex import pex_warnings
+from pex.common import environment_as
 from pex.compatibility import PY2
 from pex.pex_warnings import PEXWarning
 from pex.typing import TYPE_CHECKING
 from pex.util import named_temporary_file
 from pex.variables import NoValueError, Variables
-from testing import environment_as
 from testing.pytest.tmp import Tempdir
 
 if TYPE_CHECKING:
